attached_assets/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pytelegrambotapi>=4.26.0",
    "python-telegram-bot>=20.0",
    "sqlalchemy>=2.0.40",
    "telegram>=0.0.1",
]
